requirements.txt
flask
Werkzeug==2.0.0
requests
lmdb
flask-compress
flufl.lock
stripe
inapppy
